Refactored classOption and threadOption to grab info trait
