+++ scala-debugger-api/src/test/scala/test/InfoTestClasses.scala
@@ -140,8 +140,6 @@ object InfoTestClasses {
   class TestGrabInfoProfile extends GrabInfoProfile {
     override def threadGroup(threadGroupReference: ThreadGroupReference): ThreadGroupInfoProfile = throwException()
     override def threadGroups: Seq[ThreadGroupInfoProfile] = throwException()
-    override def threadOption(threadId: Long): Option[ThreadInfoProfile] = throwException()
-    override def classOption(name: String): Option[ReferenceTypeInfoProfile] = throwException()
     override def `object`(threadReference: ThreadReference, objectReference: ObjectReference): ObjectInfoProfile = throwException()
     override def thread(threadReference: ThreadReference): ThreadInfoProfile = throwException()
     override def classes: Seq[ReferenceTypeInfoProfile] = throwException()
